+++ program/js/editor.js
@@ -151,6 +151,16 @@ function rcube_text_editor(config, id)
     ed.on('keypress', function() {
       rcmail.compose_type_activity++;
     });
+    ed.on('PastePostProcess', function(e) {
+      $(e.node).find('img').each(function() {
+        // Need an attribute to find the image element after paste
+        var id = 'i' + Date.now();
+        this.setAttribute('data-img-id', id);
+        // Replace the 'src' with data: URI after the image has been loaded
+        // This way if we fetch the image again the browser will fetch it from cache
+        this.onload = function() { this.onload = null; ref.replace_img_src(this.src, id) };
+      });
+    });
     // make links open on shift-click
     ed.on('click', function(e) {
       var link = $(e.target).closest('a');
@@ -864,5 +874,32 @@ function rcube_text_editor(config, id)
     form.append(file).append($('<input>').attr({type: 'hidden', name: '_token', value: rcmail.env.request_token}));
 
     return wrapper.append(form);
+  };
+
+  // Replace pasted image src location with data: URI
+  this.replace_img_src = function(src, id)
+  {
+    if (!window.fetch || !src || !src.match(/^https?:\/\//)) {
+      return;
+    }
+
+    fetch(src).then(function(response) {
+      response.arrayBuffer().then(function(body) {
+        var url, ctype = response.headers.get('content-type'),
+          u8Buf = new Uint8Array(body),
+          latinBuf = '';
+
+        u8Buf.forEach(function(byte) { latinBuf += String.fromCharCode(byte); });
+
+        if (ctype && latinBuf) {
+          url = 'data:' + ctype + ';base64,' + btoa(latinBuf);
+          $(ref.editor.getBody()).find('img[data-img-id="' + id + '"]').attr({
+            src: url,
+            'data-mce-src': url,
+            'data-img-id': null
+          });
+        }
+      });
+    });
   };
 }